requirements.txt
streamlitpandasgoogle-api-python-clientgoogle-auth-httplib2google-auth-oauthlib